requirements.txt
asgiref==3.9.1
Django==5.2.6
djangorestframework==3.16.1
sqlparse==0.5.3
pytest==8.4.2
pytest-django==4.11.1
